Agregando intermedias y modificando controladores
modificando relaciones de los controladores

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use App\Tipo_usuario;
 use App\Http\Requests\UserRequest;
 
 class UserController extends Controller
@@ -33,6 +34,16 @@ class UserController extends Controller
 
         $users->save();
         return response()->json($users);
+
+
+
+    }
+
+
+    public function mostrarReservasUsuario($id){
+
+
+
     }
 
 
@@ -58,14 +69,21 @@ class UserController extends Controller
     public function update(Request $request, $id)//actualiza uno en especifico
     {
 
-        $users = User::find($id);
-        $users->nombre = $request->input('nombre');
-        $users->apellido = $request->input('apellido');
-        $users->correo = $request->input('correo');
-        $users->contrasena = $request->input('contrasena');
-        $users->id_tipo_usuario = $request->input('id_tipo_usuario');
-        $users->save();
-        return response()->json($users);
+        if($users = User::find($id)){
+            $users->nombre = $request->input('nombre');
+            $users->apellido = $request->input('apellido');
+            $users->correo = $request->input('correo');
+            $users->contrasena = $request->input('contrasena');
+            $users->id_tipo_usuario = $request->input('id_tipo_usuario');
+            $users->save();
+            return response()->json($users);
+
+        }else{
+            return "No se encuentra usuario para ser modificado.";
+
+        }
+
+
 
     }
 
@@ -73,10 +91,12 @@ class UserController extends Controller
 
     public function destroy($id)//Elimina uno en especifico
     {
-        $users = User::find($id);
-        $users ->delete();
-        return "Eliminado!";
-
+        if($users = User::find($id)) {
+            $users->delete();
+            return "Eliminado!";
+        }else{
+            return "No se pudo eliminar usuario.";
+        }
 
     }
 }
